fix: #222 편집 페이지 버그 수정 및 진행방식 데이터 네이밍 변경
