+++ src/components/recruit/create/detailedInformation/DetailedInformation.tsx
@@ -22,6 +22,8 @@ const DetailedInformation = () => {
 		setFormData({ ...formData, content: contents });
 	};
 
+	//console.log(formData.content);
+
 	return (
 		<S.DetailedInformation>
 			<section className='container-details'>
