+++ src/constant/index.ts
@@ -1,5 +1,13 @@
 import { INPUT_VALIDATION, TEXTAREA_VALIDATION } from './validation';
 import { SAFE_DEFAULT_VALUE } from './textEditor';
 import { PORTFOLIO_EDITOR_TEMPLATE } from './editorTemplate';
+import { INIT_FORM_DATA, INIT_FORM_DATA_VALID } from './initFormData';
 
-export { INPUT_VALIDATION, TEXTAREA_VALIDATION, SAFE_DEFAULT_VALUE, PORTFOLIO_EDITOR_TEMPLATE };
+export {
+	INPUT_VALIDATION,
+	TEXTAREA_VALIDATION,
+	SAFE_DEFAULT_VALUE,
+	PORTFOLIO_EDITOR_TEMPLATE,
+	INIT_FORM_DATA,
+	INIT_FORM_DATA_VALID,
+};
