+++ src/utils/index.ts
@@ -11,7 +11,6 @@ import zipFile from './zipFile';
 import unzipFile from './unzipFile';
 import isNotNumber from './isNotNumber';
 import calculateDate from './calculateDate';
-import resetFormData from './resetFormData';
 import throttle from './throttle';
 import addClassToEmptyPTags from './addClassToEmptyPTags';
 
@@ -30,7 +29,6 @@ export {
 	unzipFile,
 	isNotNumber,
 	calculateDate,
-	resetFormData,
 	throttle,
 	addClassToEmptyPTags,
 };
